requirements.txt
amqp==2.3.2
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-test-plus==1.1.1
idna==2.8
kombu==4.2.2
lxml==4.2.5
model-mommy==1.6.0
more-itertools==4.3.0
numpy==1.15.4
pandas==0.23.4
pathlib2==2.3.3
pluggy==0.8.0
py==1.7.0
pytest==4.0.2
pytest-django==3.4.4
python-dateutil==2.7.5
pytrends==4.4.0
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.1.4
